requirements.txt
fastapi==0.91.0
uvicorn==0.20.0
SQLAlchemy==2.0.3
SQLAlchemy-Utils==0.40.0
pydantic==1.10.4
psycopg2-binary==2.9.5
email-validator==1.3.1
bcrypt==4.0.1
alembic==1.9.3
django-decouple==2.1